Added segment format converter to support conversion to single file index(v3) format.
The segment format converter supports converting v1/v2 formats to v3.
Server will convert the segment format if the configuration enables it and
if an already converted segment does not exist on disk. Configuration is
supported at server and table level. Table configuration overrides server
configuration. Missing/default configuration value is V1 format. After
conversion, new directory is created as subdirectory of an existing segment
by name v3/.

Testing: Unit tests that verify format conversion, loading and ensuring
that reverting configuration allows us to reload the same segment.

Performance: Linear increase to around 2 seconds for 1.8GB segment

RB=703222
G=pinot-dev-reviewers
R=kgopalak,raringun,ssubrama
A=ssubrama

+++ pinot-core/src/test/java/com/linkedin/pinot/core/segment/index/loader/SegmentPreProcessorTest.java
@@ -97,7 +97,9 @@ public class SegmentPreProcessorTest {
   @AfterClass
   public void tearDownClass()
       throws Exception {
-    FileUtils.deleteQuietly(INDEX_DIR);
+    if (INDEX_DIR != null) {
+      FileUtils.deleteQuietly(INDEX_DIR);
+    }
   }
 
   @Test
